ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch==2.1.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
nltk==3.8.1